requirements.txt
djangorestframework
Django==2.2.5
django-redis==4.11.0
